requirements.txt
langchain 
langchain-community 
langchain-openai 
langchain-experimental 
tiktoken  
pypdf
nest_asyncio 
llama-index 
llama-parse
llama-index-readers-file
python-dotenv
langchain_cohere
fastapi
fastapi[standard] 
uvicorn
python-multipart
langgraph
neo4j
pydantic
langchain-aws 
langchain-google-community
pydrive
office365-rest-python-client
python-docx
pillow
beautifulsoup4
PyPDF2
unstructured[pdf]
unstructured
azure-storage-blob
pytesseract  
office365-rest-python-client  
python-dotenv
python-docx
pillow
requests 
beautifulsoup4 
python-docx  
unstructured[pdf]
unstructured[all] 
nltk
python-frontmatter
langchain-community
pytest
PyMuPDF
cairosvg
faiss-cpu
pinecone
langchain_chroma
langchain_pinecone
